[Signal Generator] Design the DDS system
The project's Direct Digital Synthesis (DDS) system is responsible for
generating the various user-programmable waveforms, as well as setting
the generated signal's frequency. The whole system is encapsulated
within a KiCAD hierarchical sheet. More details on the working operation
of the DDS system may be found in the (future) documentation of this
project...soonTM