[Signal Generator] Layout microcontroller and allocate its pins
The microcontroller is the brains of the function generator system - it
is responsible for controlling the DDS signal generator system, as well
as the LCD module which provides the user interface. The specific
microcontroller used in this project is the ATmega2561-16AU from the
megaAVR uC family. The reasons I have picked this particular part are
the following:
- I am familiar with the AVR uC structure;
- it provides more SRAM memory (8 kB) than other MCUs of the same family;
- it has fewer pins (and therefore peripherals) than the ATmega2560.
Only a fraction of the available peripheral pins of the ATmega2561 have
been utilized in the current design. The remaining port pins (which are
internally connected to useful peripherals such as JTAG, XMEM, etc.)
are connected to pin header connectors and are meant for
upgrading/modifying the system as a whole by including additional
external circuitry. The ATmega2561's SPI port is also connected to a
similar pin header connector which allows the uC to play the role of an
SPI slave device.