[Signal Generator] Power the LCD module with the +5VD rail
Previously, the LCD module was powered via the +3.3VD rail. The module itself
consumes about 90mA and the level shifters may consume  >= 50mA in worst
case scenarios. Looking at the module's (MSP3218) schematic, it can be
observed that the VCC line is connected to the XC6206 +3.3V LDO voltage
regulator. Therefore, I have decided to power the module from the +5VD
supply rail which is external to the board (as opposed to the +3.3VD
rail which resides entirely on the digital function generator's PCB).