[Signal Generator] Power LCD BL control digital POT from +5VD
The MCP4551 BL control POT is now powered from +5VD. Its terminals are
still connected to +3.3VD and GNDD.

Additionally, a jumper has been added which shorts the +3.3VD rail to
the LCD BL control on the LCD module (MSP3218). If this connection is
shorted, the LCD will always be on at full brightness.