[Signal Generator] Minor changes to output amplifier
1) The 2.2nF capacitor which created a real pole low-pass filter has
been removed. This is because its cutoff frequency was about 1.4MHz.
Whilst the generated function as a whole of the digital function
generator may be considered to have a frequency of 1MHz, the actual
frequency of the generated signal is in the tens of MHz (e.g.: 64MHz for
1MHz sine wave with 64 unique points).

2) The +2.5V voltage reference of the waveform amplification output stage has
been replaced by the AZ431-A. This is because the latter is much more
widespread and cheap when compared to the LM4040.