[Signal Generator] Layout essential DDS components on PCB
The most essential components of the DDS system (SRAM, DAC, oscillator,
first 12-bit counter) have been laid out on the PCB. They have also been
connected with copper tracks of appropriate width. The reason for this
early layout and interconnection of these elements is that they are to
work with very high speed digital signals.