[Signal Generator] Rework output circuitry's waveform amplification stage
The previous design of the waveform amplification stage which utilized
a digital potentiometer to regulate gain control was flawed due to the
latter. The digital potentiometer's bandwidth is 2MHz which does not
fulfill the requirements of the circuit - the output stages may need to
accommodate for signals with frequencies in the tens of MHz.

This commit proposes a new design which uses JFET transistors as voltage
controlled variable resistors. In case this new design is unsuccessful,
the gain of the waveform amplification output stage may be regulated via
a mechanical potentiometer.